backend/requirements.txt
flask
flask-cors
uvicorn
supabase
python-dotenv